chore: Add click event listeners to picker links
In this commit, we added event listeners to all Loremipsum links using JavaScript. The code gets all the links with the #picker a selector and adds a click event listener to each of them. When a link is clicked, it prevents the default behavior (i.e., navigating to a new page) and gets the corresponding main-content section using its index. Finally, it smoothly scrolls the section into view using the scrollIntoView method. This allows for a better user experience when navigating through the page.